changing default auth + post for auth

+++ routes/web.php
@@ -30,19 +30,42 @@ use App\Http\Controllers\ContactController;
 // });
 
 
+use Laravel\Fortify\Fortify;
 
+Fortify::verifyEmailView(function () {
+    return view('auth.verify-email');
+});
+Fortify::loginView(function () {
+    return view('auth.login');
+});
+Fortify::registerView(function () {
+    return view('auth.register');
+});
+Fortify::resetPasswordView(function () {
+    return view('auth.passwords.reset');
+});
+Fortify::confirmPasswordView(function () {
+    return view('auth.passwords.confirm');
+});
+Fortify::requestPasswordResetLinkView(function () {
+    return view('auth.passwords.email');
+});
+
+Route::get('/home', function(){
+    return view('blog.home');
+})->name('home');
 
 
 Route::get('/posts', [PostController::class, 'index']);
 Route::get('/', [PostController::class, 'index']);
 
-Route::get('/posts/create', [PostController::class, 'create']);
+Route::get('/posts/create', [PostController::class, 'create'])->middleware('auth');
 
 Route::get('/posts/{post}', [PostController::class, 'show']);
-Route::post('/posts', [PostController::class, 'store']);
-Route::get('/posts/{post}/edit', [PostController::class, 'edit']);
-Route::put('/posts/{post}', [PostController::class, 'update']);
-Route::delete('/posts/{post}', [PostController::class, 'destroy']);
+Route::post('/posts', [PostController::class, 'store'])->middleware('auth');
+Route::get('/posts/{post}/edit', [PostController::class, 'edit'])->middleware('auth');
+Route::put('/posts/{post}', [PostController::class, 'update'])->middleware('auth');
+Route::delete('/posts/{post}', [PostController::class, 'destroy'])->middleware('auth');
 
 Route::get('/about', function(){
     return view('blog.about');
